+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Upgrade to psycopg2 2.8.3
 - Add Go usage example
 - Send monthly reports on 1st of every month, not randomly during the month
+- Signup form sets the "auto-login" cookie to avoid an extra click during first login
 
 ### Bug Fixes
 - Prevent double-clicking the submit button in signup form